+++ src/main/java/org/khee/kheetun/client/TunnelClientListener.java
@@ -4,6 +4,6 @@ import org.khee.kheetun.client.config.Tunnel;
 
 public interface TunnelClientListener {
     
-    public void TunnelClientConnection( boolean connected );
+    public void TunnelClientConnection( boolean connected, String error );
     public void TunnelClientTunnelStatus( Tunnel tunnel );
 }